requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
pandas>=2.2,<3.1
pyarrow>=16,<24
joblib>=1.4,<2.0
scikit-learn>=1.5,<1.9
pydantic>=2.8,<3.0
httpx>=0.28,<1.0
pytest>=9.0,<10.0
streamlit>=1.45,<2.0
python-docx>=1.2,<2.0
